On branch main Changes to be committed: 	modified:   CMakeLists.txt 	new file:   README.md 	modified:   src/CPU/solver.cpp 	modified:   src/GPU/include/solver.cuh 	modified:   src/GPU/kernel.cu 	modified:   src/GPU/solver.cu 	modified:   test/Plot.py

+++ src/GPU/include/solver.cuh
@@ -13,32 +13,32 @@ inline int Block;
 
 /*define some variables which respond to the physical status */
 struct status {
-  double x;   /*position*/
-  double rou; /*density;*/
-  double u;   /*velocity;*/
-  double p;   /*pressure;*/
-  double E;   /*total energy;*/
-  double T;   /*tmpeture*/
-  double c;   /*the speed of sound*/
+  float x;   /*position*/
+  float rou; /*density;*/
+  float u;   /*velocity;*/
+  float p;   /*pressure;*/
+  float E;   /*total energy;*/
+  float T;   /*tmpeture*/
+  float c;   /*the speed of sound*/
 };
 struct gpuStatus {
-  double *x = nullptr;
-  double *rou = nullptr;
-  double *u = nullptr;
-  double *p = nullptr;
-  double *E = nullptr;
-  double *T = nullptr;
-  double *c = nullptr;
+  float *x = nullptr;
+  float *rou = nullptr;
+  float *u = nullptr;
+  float *p = nullptr;
+  float *E = nullptr;
+  float *T = nullptr;
+  float *c = nullptr;
 };
 struct gpuU {
-  double *u1 = nullptr;
-  double *u2 = nullptr;
-  double *u3 = nullptr;
+  float *u1 = nullptr;
+  float *u2 = nullptr;
+  float *u3 = nullptr;
 };
 struct Vector {
-  double rou_r;
-  double rou_u;
-  double rou_e;
+  float rou_r;
+  float rou_u;
+  float rou_e;
 };
 /*help to specific the physical meaning of U and F for different index*/
 typedef enum {
@@ -49,10 +49,10 @@ typedef enum {
 //------------------------------------------------------//
 class Euler {
 private:
-  double R0 = 287.0;
-  double gamma = 1.4;
+  float R0 = 287.0;
+  float gamma = 1.4;
   int size;
-  double dt, dx, t;
+  float dt, dx, t;
   status *data = nullptr;
   Vector *U = nullptr;
   /*gpu*/
@@ -60,12 +60,12 @@ private:
   gpuU gpu_U;
   cudaDeviceProp device_prop;
   cudaEvent_t kernel_start, kernel_end;
-  double runtime = 0.0;
+  float runtime = 0.0;
   /*update data*/
   void refresh();
 
 public:
-  Euler(int size, double t);
+  Euler(int size, float t);
   ~Euler();
   void Runge_Kutta();
   void solver();
